backend/requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
prefect
mlflow

